Split dark/light theme switch into separate component for modularity
